requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.8.2
python-pptx==0.6.23
Pillow==10.4.0
tenacity==8.5.0